// src/app/models/userlist.ts
export class UserList {
    userId: number;
    name: string;
    email: string;
    eid: string;
    profileImage: string;
    active: string;
    surname:string;
    adress:string;
    city:string;
    countryCode :string;
    country :String;
    postalcode :string;
    phone:string;
    phone2 :string;
    partnerId: number;
    loginType: number;
}